re-organize class name
