+++ FluentHub/Views/AppSettings/AboutPage.xaml
@@ -1,0 +1,75 @@
+﻿<Page
+    x:Class="FluentHub.Views.AppSettings.AboutPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="using:FluentHub.Views.AppSettings"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
+    xmlns:uc="using:FluentHub.UserControls"
+    xmlns:vm="using:FluentHub.ViewModels.AppSettings"
+    mc:Ignorable="d">
+
+    <Page.DataContext>
+        <vm:AboutViewModel x:Name="ViewModel" />
+    </Page.DataContext>
+
+    <ScrollViewer>
+        <Grid Margin="0,0,16,0">
+            <StackPanel
+                HorizontalAlignment="Stretch"
+                VerticalAlignment="Stretch"
+                Spacing="4">
+                <uc:SettingsBlockControl
+                    Title="FluentHub Dev"
+                    Description="{x:Bind ViewModel.Version}"
+                    IsClickable="False">
+                    <uc:SettingsBlockControl.Icon>
+                        <muxc:ImageIcon
+                            Height="36"
+                            Margin="-6,0"
+                            Source="/Assets/AppTiles/DefaultLogo.png" />
+                    </uc:SettingsBlockControl.Icon>
+                    <Button
+                        Width="36"
+                        Height="32"
+                        Padding="0"
+                        ToolTipService.ToolTip="Copy">
+                        <FontIcon FontSize="14" Glyph="&#xE8C8;" />
+                    </Button>
+                </uc:SettingsBlockControl>
+                <uc:SettingsBlockControl Title="Useful links">
+                    <uc:SettingsBlockControl.Icon>
+                        <FontIcon Glyph="&#xE71B;" />
+                    </uc:SettingsBlockControl.Icon>
+                    <uc:SettingsBlockControl.ExpandableContent>
+                        <Grid Margin="0,12,0,12" HorizontalAlignment="Center">
+                            <Grid.ColumnDefinitions>
+                                <ColumnDefinition Width="Auto" />
+                                <ColumnDefinition Width="Auto" />
+                                <ColumnDefinition Width="Auto" />
+                                <ColumnDefinition Width="Auto" />
+                            </Grid.ColumnDefinitions>
+                            <HyperlinkButton
+                                Grid.Column="0"
+                                Content="Source code"
+                                NavigateUri="https://github.com/fluenthub-uwp/FluentHub" />
+                            <HyperlinkButton Grid.Column="1" Content="Privacy policy" />
+                            <HyperlinkButton Grid.Column="2" Content="License" />
+                            <HyperlinkButton Grid.Column="3" Content="Third party notices" />
+                        </Grid>
+                    </uc:SettingsBlockControl.ExpandableContent>
+                </uc:SettingsBlockControl>
+                <uc:SettingsBlockControl
+                    Title="Open app log location"
+                    IsClickable="True">
+                    <uc:SettingsBlockControl.Icon>
+                        <FontIcon Glyph="&#xe838;" />
+                    </uc:SettingsBlockControl.Icon>
+                    <FontIcon FontSize="16" Glyph="&#xE72D;" />
+                </uc:SettingsBlockControl>
+            </StackPanel>
+        </Grid>
+    </ScrollViewer>
+</Page>
+
