+++ FluentHub/Views/AppSettings/AboutPage.xaml.cs
@@ -13,16 +13,11 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-// La plantilla de elemento Página en blanco está documentada en https://go.microsoft.com/fwlink/?LinkId=234238
-
-namespace FluentHub.Views.AppSettingsPages
+namespace FluentHub.Views.AppSettings
 {
-    /// <summary>
-    /// Una página vacía que se puede usar de forma independiente o a la que se puede navegar dentro de un objeto Frame.
-    /// </summary>
-    public sealed partial class AboutUsPage : Page
+    public sealed partial class AboutPage : Page
     {
-        public AboutUsPage()
+        public AboutPage()
         {
             this.InitializeComponent();
         }
