+++ FluentHub/Views/AppSettings/AppearancePage.xaml
@@ -1,18 +1,18 @@
 ﻿<Page
-    x:Class="FluentHub.Views.AppSettingsPages.GeneralSettingsPage"
+    x:Class="FluentHub.Views.AppSettings.AppearancePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:FluentHub.Views.AppSettingsPages"
+    xmlns:local="using:FluentHub.Views.AppSettings"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     xmlns:uc="using:FluentHub.UserControls"
-    xmlns:vm="using:FluentHub.ViewModels.AppSettingsPages"
+    xmlns:vm="using:FluentHub.ViewModels.AppSettings"
     Loaded="Page_Loaded"
     mc:Ignorable="d">
 
     <Page.DataContext>
-        <vm:GeneralPageViewModel x:Name="ViewModel" />
+        <vm:AppearanceViewModel x:Name="ViewModel" />
     </Page.DataContext>
     <Grid>
         <Grid.ColumnDefinitions>
@@ -20,24 +20,21 @@
         </Grid.ColumnDefinitions>
         <StackPanel Grid.Column="0" Spacing="4">
             <uc:SettingsBlockControl
-                Title="Theme"
-                Description="Choose the theme that you want to use in the app."
+                Title="App Theme"
+                Description="Select which app theme to display"
                 IsClickable="False">
                 <uc:SettingsBlockControl.Icon>
                     <FontIcon FontFamily="{StaticResource Fluent Icons}" Glyph="&#xE771;" />
                 </uc:SettingsBlockControl.Icon>
-                <ComboBox
-                    x:Name="themeComboBox"
-                    PlaceholderText="Choose your theme"
-                    SelectionChanged="ComboBox_SelectionChanged">
-                    <ComboBoxItem Content="System Default" Tag="Default" />
+                <ComboBox x:Name="themeComboBox" SelectionChanged="ComboBox_SelectionChanged">
                     <ComboBoxItem Content="Light" Tag="Light" />
                     <ComboBoxItem Content="Dark" Tag="Dark" />
+                    <ComboBoxItem Content="Use system settings" Tag="Default" />
                 </ComboBox>
             </uc:SettingsBlockControl>
             <uc:SettingsBlockControl
                 Title="Language"
-                Description="Choose the language that you want to use in the app."
+                Description="Select language which you want use"
                 IsClickable="False">
                 <uc:SettingsBlockControl.Icon>
                     <FontIcon FontFamily="{StaticResource Fluent Icons}" Glyph="&#xF2B7;" />
