+++ FluentHub/Views/AppSettings/MainSettingsPage.xaml
@@ -1,9 +1,9 @@
 ﻿<Page
-    x:Class="FluentHub.Views.AppSettingsPages.GeneralPage"
+    x:Class="FluentHub.Views.AppSettings.MainSettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:FluentHub.Views.AppSettingsPages"
+    xmlns:local="using:FluentHub.Views.AppSettings"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     mc:Ignorable="d">
