--- FluentHub/Views/AppSettingsPages/AboutUsPage.xaml
@@ -1,47 +1,0 @@
-﻿<Page
-    x:Class="FluentHub.Views.AppSettingsPages.AboutUsPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:FluentHub.Views.AppSettingsPages"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:uc="using:FluentHub.UserControls"
-    mc:Ignorable="d">
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-
-        <Image
-            Grid.Row="0"
-            Width="300"
-            Margin="0,10,0,0"
-            HorizontalAlignment="Left"
-            Source="/Assets/FluentHub.png" />
-
-        <Grid Grid.Row="1">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-
-            <StackPanel Grid.Column="0" Margin="0,23,0,0">
-                <uc:SettingsBlockControl
-                    Title="About us"
-                    Description="FluentHub is an amazing app that at the moment is in Dev stage"
-                    IsClickable="False">
-                    <uc:SettingsBlockControl.Icon>
-                        <FontIcon FontFamily="{StaticResource Fluent Icons}" Glyph="&#xE07F;" />
-                    </uc:SettingsBlockControl.Icon>
-                </uc:SettingsBlockControl>
-            </StackPanel>
-
-
-        </Grid>
-
-
-
-    </Grid>
-</Page>
-
